[REFACTOR] : SidedishCategoryService를 이용하여 중복 제거
