+++ be/sidedish-be/src/main/java/com/codesquad/sidedish/web/sidedish/service/SidedishCategoryService.java
@@ -2,34 +2,35 @@ package com.codesquad.sidedish.web.sidedish.service;
 
 import com.codesquad.sidedish.web.sidedish.DTO.ItemDTO;
 import com.codesquad.sidedish.web.sidedish.domain.Image;
+import com.codesquad.sidedish.web.sidedish.domain.SidedishCategory;
 import com.codesquad.sidedish.web.sidedish.repository.ImageRepository;
 import com.codesquad.sidedish.web.sidedish.repository.SidedishCategoryRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
-public class SideSidedishesService {
+public class SidedishCategoryService {
     SidedishCategoryRepository sidedishCategoryRepository;
     ImageRepository imageRepository;
 
-    public SideSidedishesService (SidedishCategoryRepository sidedishCategoryRepository, ImageRepository imageRepository) {
+    public SidedishCategoryService(SidedishCategoryRepository sidedishCategoryRepository, ImageRepository imageRepository) {
         this.sidedishCategoryRepository = sidedishCategoryRepository;
         this.imageRepository = imageRepository;
     }
 
-    public List<ItemDTO> getSideSidedishList() {
-        Map<Long, Image> images = new HashMap<>();
+    public List<ItemDTO> readSideSidedishesBy(String categoryName) {
+        Map<Long, Image> images = imageRepository.findAll().stream()
+                .collect(Collectors.toMap(Image::getId, image -> image));
 
-        for(Image i : imageRepository.findAll()) {
-            images.put(i.getId(), i);
-        }
-
-        return sidedishCategoryRepository.findByNameIs("밑반찬").getSidedishes()
-                .stream().map(s -> ItemDTO.of(s, images))
+        return readBy(categoryName).sidedishStream()
+                .map(sidedish -> ItemDTO.of(sidedish, images))
                 .collect(Collectors.toList());
+    }
+
+    private SidedishCategory readBy(String categoryName) {
+        return sidedishCategoryRepository.findByNameIs(categoryName);
     }
 }
